% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF_model_lm.R
\name{CDF_model_lm}
\alias{CDF_model_lm}
\title{Creates the probability predictive distribution of an lm() model}
\usage{
CDF_model_lm(x_cal, model)
}
\arguments{
\item{x_cal}{Observation vector/matrix of the calibration set}

\item{model}{A lm() model}
}
\value{
A cumulative predictive distribution function of a normal distribution
}
\description{
Creates the probability predictive distribution of an lm() model
}
\examples{
n <- 100000

# generationg heterocedastic data
mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_test <- x[1:80000]
y_test <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_test ~ x_test)

cdf <- CDF_model_lm(x_cal=x_cal, model= model)


}
