% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_global_pit.R
\name{gg_global_pit}
\alias{gg_global_pit}
\title{Plot to diagnose calibration bases on  p_values}
\usage{
gg_global_pit(pit)
}
\arguments{
\item{pit}{vector of pit values}
}
\value{
a ggplot histogram graph
}
\description{
Plot to diagnose calibration bases on  p_values
}
\examples{

n <- 100000

# generating heterocedastic data
mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_test <- x[1:80000]
y_test <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_test ~ x_test)

cdf <- CDF_model_lm(x_cal=x_cal, model= model)

pit <- PIT_global_lm( y_cal, cdf)

gg_global_pit(pit)




}
