% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_local_pit.R
\name{gg_local_pit}
\alias{gg_local_pit}
\title{Density grsph for local PIT}
\usage{
gg_local_pit(
  pit_local,
  alpha = 0.4,
  linewidth = 1,
  pal = "Set2",
  facet = F,
  ...
)
}
\arguments{
\item{pit_local}{Data frame with two collumn names "part" and "pit"}

\item{alpha}{double 0-1 to indicate transparency of fill}

\item{linewidth}{integer linewidth of density line}

\item{pal}{a chosen RBrewer color pallete}

\item{facet}{Logical}

\item{...}{Other parameters for ggplot()}
}
\value{
A ggplot
}
\description{
Density grsph for local PIT
}
\examples{
n <- 100000
mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
 30*x1
}

x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_test <- x[1:80000]
y_test <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_test ~ x_test)

pit_local <- PIT_local_lm(xcal=x_cal, ycal=y_cal, mod=model)

gg_local_pit(pit_local)
}
