% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_PIT_global.R
\name{gg_PIT_global}
\alias{gg_PIT_global}
\title{Plot to diagnose calibration bases on  p_values}
\usage{
gg_PIT_global(pit, ...)
}
\arguments{
\item{pit}{vector of pit values}

\item{...}{other static parameters passed to geom_density()}
}
\value{
a ggplot density graph
}
\description{
A function based on ggplot2 to observe the global the density of PIT-values.
Layers can be edited like in http://cran.nexr.com/web/packages/ggpmisc/vignettes/user-guide-4.html.
}
\examples{

n <- 100000

# generating heterocedastic data
mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_train <- x[1:80000]
y_train <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_train ~ x_train)
y_hat <- predict(model, newdata=data.frame(x_train=x_cal))
MSE <- (summary(model)$sigma)^2
pit <- PIT_global(ycal=y_cal, yhat=y_hat, mse=MSE)

gg_PIT_global(pit)




}
