% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_QQ_local.R
\name{gg_QQ_local}
\alias{gg_QQ_local}
\title{QQPlot diagnose calibration locally}
\usage{
gg_QQ_local(
  pit_local,
  psz = 0.01,
  abline = "black",
  pal = "Set1",
  facet = FALSE,
  ...
)
}
\arguments{
\item{pit_local}{A dataframe obtained from PIT_local_lm}

\item{psz}{double that indicates size of the points that compose the lines.
Default is 0.001}

\item{abline}{Color of hotizontal line that indicates density 1. Default is"red"}

\item{pal}{Palette name from RColor Brewer. Default is "Set2'}

\item{facet}{logical value in case separate visualization is prefered. Default is F}

\item{...}{Other parameters to pass ggplot}
}
\value{
a ggplot graph
}
\description{
ggplot to visualize predicted vs empirical cumulative distributions locally
}
\examples{

n <- 100000

mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_train <- x[1:80000]
y_train <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_train ~ x_train)
y_hat <- predict(model, newdata=data.frame(x_train=x_cal))
MSE <- (summary(model)$sigma)^2
pit_local <- PIT_local(xcal = x_cal, ycal=y_cal, yhat=y_hat, mse=MSE)

gg_QQ_local(pit_local)
gg_QQ_local(pit_local, facet=TRUE)

}
