% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT_local_lm.R
\name{PIT_local_lm}
\alias{PIT_local_lm}
\title{Creates pit values for k partition of the calibration set}
\usage{
PIT_local_lm(
  xcal,
  ycal,
  clusters = 5,
  n_neighboor = 1000,
  fx = CDF_model_lm,
  pit = PIT_values_lm,
  mod
)
}
\arguments{
\item{xcal}{Covariates (features) vector/matrix of the calibration set}

\item{ycal}{Output/Observations vector of the calibration set}

\item{clusters}{Number of local clusters one whats to evaluate}

\item{n_neighboor}{Number of neighbors for the mean of each cluster}

\item{fx}{Predictive Cumulative Distribution of the fitted lm() model.}

\item{pit}{Function that returns pit values of a lm() model.
correspond to the pit values for each partition.}

\item{mod}{A lm() model.}
}
\value{
Data Frame of pit-values in different partitions.
}
\description{
Creates pit values for k partition of the calibration set
}
\examples{
n <- 100000

mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_test <- x[1:80000]
y_test <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_test ~ x_test)

PIT_local_lm(xcal=x_cal, ycal=y_cal, mod=model)



}
