% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT_values_lm.R
\name{PIT_values_lm}
\alias{PIT_values_lm}
\title{Obtain pit-values for an fitted lm() model}
\usage{
PIT_values_lm(y_cal, pdf)
}
\arguments{
\item{y_cal}{a vector of outputs of calibration set}

\item{pdf}{A predective cumulative function}
}
\value{
a vector of pit-values
}
\description{
Obtain pit-values for an fitted lm() model
}
\examples{
n <- 100000

# generationg heterocedastic data
mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_test <- x[1:80000]
y_test <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_test ~ x_test)

pdf <- PDF_model_lm(x_cal=x_cal, model=model)

PIT_values_lm( y_cal, pdf)



}
