% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate.R
\name{recal_exact}
\alias{recal_exact}
\title{Title}
\usage{
recal_exact(
  x_cal,
  x_test,
  y_cal_hat,
  p_values,
  mse_val,
  n_neighboors,
  epsilon = 0
)
}
\arguments{
\item{x_cal}{features/covariates from validation set}

\item{x_test}{features/covariates from test set}

\item{mse_val}{mse from the fitted model, assuming normality}

\item{epsilon}{}

\item{y_cal_pred}{predicted observations from the predictive distribution}
}
\description{
Title
}
\examples{
n <- 100000
#generationg heterocedastic data
mu <- function(x1){
 10 + 5*x1^2
}
sigma_v <- function(x1){
 30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_test <- x[1:80000]
y_test <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_test ~ x_test)
y_cal_hat <- predict(model, data.frame(x_test=x_cal))

cdf <- CDF_model_lm(x_cal=x_cal, model= model)

MSE <- summary(model)$sigma

pit <- PIT_local(xcal=x_cal, y_cal=y_cal, xnew=y_test, y_hat=y_pred,  MSE=MSE)
}
