% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_QQ_global.R
\name{gg_QQ_global}
\alias{gg_QQ_global}
\title{QQPlot diagnose calibration globally}
\usage{
gg_QQ_global(pit, ycal, yhat, mse)
}
\arguments{
\item{pit}{vector of global PIT-values}

\item{ycal}{vector of y calibration set}

\item{yhat}{vector of predicted y on calibration set}

\item{mse}{Mean Squared Error from fitted model}
}
\value{
a ggplot point graph
}
\description{
ggplot to visualize predicted vs empirical cumulative distributions
}
\examples{

n <- 100000

# generating heterocedastic data
mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}


x <- runif(n, 2, 20)
y <- rnorm(n, mu(x), sigma_v(x))

x_train <- x[1:80000]
y_train <- y[1:80000]

x_cal <- x[80001:100000]
y_cal <- y[80001:100000]

model <- lm(y_train ~ x_train)
y_hat <- predict(model, newdata=data.frame(x_test=x_cal))
MSE <- (summary(model)$sigma)^2

pit <- PIT_global( y_cal, y_hat,  MSE)


gg_QQ_global(pit,y_cal, y_hat, MSE)


}
