% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT_local.R
\name{PIT_local}
\alias{PIT_local}
\title{Title}
\usage{
PIT_local(
  xcal,
  ycal,
  xnew,
  yhat,
  clusters = 6,
  n_neighboor = 1000,
  pit = PIT_global,
  MSE
)
}
\arguments{
\item{MSE}{}
}
\description{
Title
}
