% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate.R
\name{recalibrate}
\alias{recalibrate}
\title{Obtain recalibrated samples of the predictive distribution}
\usage{
recalibrate(
  xcal,
  xnew,
  yhat,
  pit_values,
  mse,
  n_neighboors = 1000,
  epsilon = 1,
  method = c("gui", "kuleshov1"),
  type = c("local", "global")
)
}
\description{
Obtain recalibrated samples of the predictive distribution
}
